requirements.txt
requests
pyperclip
pystray
pillow
tkinter